improvements to gitconfig
new options:
- useful defaults for `git status`, `stash`, and `submodule`
- aliases for above adjusted to respect and use new defaults

alias "rework":

Aliases now have a pneumonic for a command, and that pneumonic is
repeated everywhere that command or its concept is. The pneumonic is at
the top of a paragraph for that command. Specializations or common flags
are then alias+flag-initials/pneumonic. Commands referencing other
commands or concepts use the appropriate pneumonic E.g. rebase is rb, so
rebase commands use rb and git pull --rebase uses rb. Before this was
not the case.